Modularizing zfinger more
Also removed debugging stuff added in earlier about checking local
availability of searches to zfinger.
